// src/types/justiceInquiry.type.ts
export type JusticeInquiryRate = {
	justiceInquiryRateId: string;
	translationItemId: string;
	translationItemName: string;
	translationItemIsActive: boolean;
	languageId: string;
	languageName: string;
	justiceInquiryId: string;
	justiceInquiryName: string;
	languageIsActive: boolean;
	isRequired: boolean;
	justiceInquiryIsActive: boolean;
	price: number | string;
};